Refactoring of database access layer
We now have a really minimal ORM inside the Sqlite layer so that all
serialization and parsing occours right at the DAO level, not sprinkled
everywhere across the code base. We have a engineering tradeoff here
where sometimes we parse too much (e.g., JSON parsing) but we do this
to avoid the added complexity of classes that perform lazy eval (which
would be a doubtful optimization at best regarding my gut feeling).