Further work on multirun printing
We collect the results, but now printing works quite differently because
the data structure is laid out so that in the regular case, we always
retrieve by multirun_id (and then select all run_ids from that). In the
current display logic, this is the other way around and requires an
awkward (!) workaround.