Rework of result saving structure
Most of the time we're not interested in failed replies. Actually, we
only care about a few failed replies (so we have something to show) but
all Pass or NoAnswer replies just occupy space in the database without
adding anything. Additionally, keeping the summary precalculated greatly
improves performance. Since we changed the DB format anyways do this
refactoring at the same time.